doc: Add some comments to get a better understanding of the code and add the program schematic to the readme.
Add some comments to the code in order to get a solid understanding of how the code works. Also
the aimed program schematic is added to the readme, which is realized by working on the
given tasks that need to be fulfilled in order to finish the traffic simulation project for the
nanodegree.

No issue addressed