feat: Create the TrafficLight class and the definition as well as a rough implementation of the MessageQueue class
Create the TrafficLight class as well as the MessageQueue class in order to fulfill the tasks FP.1, FP.2 and FP.3.
The current thread safe message passing is meat to be with the methods WaitForGreen() of TrafficLight by
incooperating the thread safe MessageQueue class that will be called by the Intersection class.

No issue addressed